pyproject.toml
[project]
name = "python1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.13",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
